move background logic to seperate files and refactor test-setup a bit to prepare for further tests
